sending the same seed using a udp server

+++ src/main/java/org/ort_rehovot/bubble_shooter/ipc/CommandFormatter.java
@@ -5,7 +5,7 @@ public class CommandFormatter {
         return "H" + ' ' + port;
     }
 
-    public static String ready(int port, int seed, String ip) {
+    public static String ready(int port, long seed, String ip) {
         return "R" + ' ' + port + ' ' + ip + ' ' + seed;
     }
